[wdm] Work on a library.d directory
