+++ subprojects/wdm/library.d/gpu.yaml
@@ -1,0 +1,26 @@
+---
+name: has_gpu_nodes
+spec:
+  configuration:
+  - instance_type
+  - instance_count
+  test:
+  - name: has_gpu_nodes
+    type: toolbox
+    spec:
+      group: nfd
+      command: has_gpu_nodes
+  install:
+  - name: add_gpu_nodes
+    type: toolbox
+    spec:
+      group: cluster
+      command: set_scale
+      args:
+      - $instance_type
+      - $instance_count
+  - name: wait_gpu_labels
+    type: toolbox
+    spec:
+      group: nfd
+      command: wait_gpu_nodes
