+++ subprojects/wdm/requirements.txt
@@ -1,0 +1,3 @@
+pydantic
+pyyaml
+fire
