+++ testing/prow/wdm/entitlement.yaml
@@ -1,0 +1,26 @@
+---
+name: has_entitlement
+spec:
+  test:
+  - name: test_entitlement
+    type: toolbox
+    spec:
+      group: entitlement
+      command: test_cluster
+      args:
+      - --no_inspect
+  install:
+  - name: deploy_entitlement
+    type: toolbox
+    configuration:
+    - entitlement_pem
+    spec:
+      group: entitlement
+      command: deploy
+      args:
+      - "$entitlement_pem"
+  - name: wait_entitlement
+    type: toolbox
+    spec:
+      group: entitlement
+      command: wait
