--- testing/wdm/gpu-operator.yml
@@ -1,134 +1,0 @@
-name: has_gpu_operator
-spec:
-  requirements:
-  - has_nfd
-  - has_gpu_nodes
-  tests:
-  - name: has_gpu_operator
-    type: shell
-    spec: oc get pod -l app.kubernetes.io/component=gpu-operator -A -oname | grep .
-  install:
-  - name: install_gpu_operator
-    type: toolbox
-    spec:
-      group: gpu_operator
-      command: deploy_from_operatorhub
-  - name: wait_gpu_deployment
-    type: toolbox
-    spec:
-      group: gpu_operator
-      command: wait_deployment
----
-name: has_nfd
-spec:
-  requirements:
-  - has_nfd_in_operatorhub
-  - has_nfd_from_operatorhub
----
-name: has_nfd_in_operatorhub
-spec:
-  tests:
-  - name: has_nfd_operatorhub
-    type: shell
-    spec: oc get packagemanifests/nfd -n openshift-marketplace > /dev/null
----
-name: has_nfd_labels
-spec:
-  tests:
-  - name: has_nfd_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: has_labels
----
-name: has_nfd_from_operatorhub
-spec:
-  tests:
-  - name: has_nfd_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: has_labels
-  install:
-  - name: install_nfd_from_operatorhub
-    type: toolbox
-    spec:
-      group: nfd_operator
-      command: deploy_from_operatorhub
-  - name: wait_nfd_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: wait_labels
----
-name: has_nfd_from_master
-spec:
-  tests:
-  - name: has_nfd_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: has_labels
-  install:
-  - name: has_nfd_from_master
-    type: toolbox
-    spec:
-      group: nfd-operator
-      command: deploy_from_commit
-      args:
-      - "${CI_IMAGE_NFD_COMMIT_CI_REPO}"
-      - "${CI_IMAGE_NFD_COMMIT_CI_REF}"
-      - image-tag="${CI_IMAGE_NFD_COMMIT_CI_IMAGE_TAG}"
-  - name: wait_nfd_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: wait_labels
----
-name: has_gpu_nodes
-spec:
-  tests:
-  - name: has_gpu_nodes
-    type: toolbox
-    spec:
-      group: nfd
-      command: has_gpu_nodes
-  install:
-  - name: add_gpu_nodes
-    type: toolbox
-    spec:
-      group: cluster
-      command: set_scale
-      arguments:
-      - ${INSTANCE_TYPE:-g4dn.xlarge}
-      - ${INSTANCE_COUNT:-1}
-  - name: wait_gpu_labels
-    type: toolbox
-    spec:
-      group: nfd
-      command: wait_gpu_nodes
----
-name: has_entitlement
-spec:
-  tests:
-  - name: test_entitlement
-    type: toolbox
-    spec:
-      group: entitlement
-      command: test_cluster
-      args:
-      - --no_inspect
-  install:
-  - name: deploy_entitlement
-    type: toolbox
-    spec:
-      group: entitlement
-      command: deploy
-      args:
-      - "${ENTITLEMENT_PEM}"
-  - name: wait_entitlement
-    type: toolbox
-    spec:
-      group: entitlement
-      command: wait
----
