+++ toolbox/wdm
@@ -1,1 +1,6 @@
-../subprojects/wdm/wdm.py
+#! /bin/bash
+
+THIS_DIR=$(dirname "$(realpath ${BASH_SOURCE[0]})")
+
+export PYTHONPATH=${THIS_DIR}/../subprojects/wdm/
+exec python3 -m wdm "$@"
